requirements.txt
allure-behave==2.13.5
allure-python-commons==2.13.5
Appium-Python-Client==1.3.0
attrs==24.2.0
behave==1.2.6
behave-html-formatter==0.9.10
certifi==2024.7.4
charset-normalizer==3.3.2
comtypes==1.4.6
idna==3.7
parse==1.20.2
parse-type==0.6.2
pluggy==1.5.0
pywin32==306
pywinauto==0.6.8
requests==2.32.3
selenium==3.141.0
six==1.16.0
urllib3==1.26.15
PyYAML~=6.0.2